requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.8

celery~=5.4.0
